+++ src/components/Widget/components/Launcher/index.js
@@ -1,11 +1,18 @@
-import React, { useContext } from 'react';
+import React, { useState, useEffect, useContext, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Map } from 'immutable';
 import ImmutablePropTypes from 'react-immutable-proptypes';
+import { usePopper } from 'react-popper';
+import Slider from 'react-slick';
+import 'slick-carousel/slick/slick.css';
+import 'slick-carousel/slick/slick-theme.css';
+
 import { MESSAGES_TYPES } from 'constants';
 import { Image, Message, Buttons } from 'messagesComponents';
 import { showTooltip as showTooltipAction } from 'actions';
+import { onRemove } from 'utils/dom';
+
 import openLauncher from 'assets/launcher_button.svg';
 import closeIcon from 'assets/clear-button-grey.svg';
 import close from 'assets/clear-button.svg';
@@ -24,15 +31,67 @@ const Launcher = ({
   unreadCount,
   displayUnreadCount,
   showTooltip,
-  lastMessage,
-  closeTooltip
+  lastMessages,
+  closeTooltip,
+  lastUserMessage,
+  domHighlight
 }) => {
   const { mainColor, assistBackgoundColor } = useContext(ThemeContext);
 
+  const [referenceElement, setReferenceElement] = useState(null);
+
+  useEffect(() => {
+    const setReference = (selector) => {
+      const reference = document.querySelectorAll(selector);
+      if (reference && reference.length === 1) {
+        onRemove(reference[0], () => setReferenceElement(null));
+        setReferenceElement(reference[0]);
+      } else {
+        setReferenceElement(null);
+      }
+    };
+    if (lastUserMessage && lastUserMessage.get('nextMessageIsTooltip')) {
+      setReference(lastUserMessage.get('nextMessageIsTooltip'));
+    } else if (domHighlight && domHighlight.get('selector')) {
+      setReference(domHighlight.get('selector'));
+    } else {
+      setReferenceElement(null);
+    }
+  }, [lastUserMessage, domHighlight]);
+
+  const [popperElement, setPopperElement] = useState(null);
+  const [arrowElement, setArrowElement] = useState(null);
+
+  const { styles, attributes } = usePopper(referenceElement, popperElement, {
+    modifiers: [
+      // The arrow padding ensures it never get on the border where it looks ugly
+      { name: 'arrow', options: { element: arrowElement, padding: 5 } },
+      {
+        name: 'preventOverflow',
+        options: {
+          padding: 15 // 0 by default
+        }
+      }
+    ],
+    placement: (domHighlight && domHighlight.get('tooltipPlacement')) || 'auto'
+  });
+
   const className = ['rw-launcher'];
+
+  const sliderSettings = {
+    dots: true,
+    infinite: false,
+    adaptiveHeight: true
+  };
+  const lastMessage = lastMessages ? lastMessages.slice(-1)[0] : new Map();
+  // This is used to distinguish bw drag and click events in the tooltips sequences.
+  const dragStatus = useRef({
+    x: 0,
+    y: 0
+  });
+
   if (isChatOpen) className.push('rw-hide-sm');
   if (fullScreenMode && isChatOpen) className.push('rw-full-screen rw-hide');
-
 
   const getComponentToRender = (message) => {
     const ComponentToRender = (() => {
@@ -54,20 +113,69 @@ const Launcher = ({
     toggle(); // open the chat if the tooltip do not know how to display the compoment
   };
 
-
-  const renderToolTip = () => (
-    <div className="rw-tooltip-body" style={{ backgroundColor: assistBackgoundColor }}>
-      <div className="rw-tooltip-close" >
-        <button onClick={(e) => { e.stopPropagation(); closeTooltip(); }}>
-          <img
-            src={closeIcon}
-            alt="close"
-          />
+  const renderTooltipContent = () => (
+    <React.Fragment>
+      <div className="rw-tooltip-close">
+        <button
+          onClick={(e) => {
+            /* stop the propagation because the popup is also a button
+            otherwise it would open the webchat when closing the tooltip */
+            e.stopPropagation();
+            closeTooltip();
+          }}
+        >
+          <img src={closeIcon} alt="close" />
         </button>
       </div>
-      <div className="rw-tooltip-response">
-        {getComponentToRender(lastMessage)}
+      {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions */}
+      <div className="rw-slider-safe-zone" onClick={e => e.stopPropagation()}>
+        <Slider {...sliderSettings}>
+          {lastMessages.map(message => (
+            // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+            <div
+              className="rw-tooltip-response"
+              onMouseDown={(event) => {
+                dragStatus.current.x = event.clientX;
+                dragStatus.current.y = event.clientY;
+              }}
+              onMouseUp={(event) => {
+                if (
+                  Math.abs(dragStatus.current.x - event.clientX) +
+                    Math.abs(dragStatus.current.y - event.clientY) <
+                  15
+                ) {
+                  toggle();
+                }
+              }}
+            >
+              {getComponentToRender(message)}
+            </div>
+          ))}
+        </Slider>
       </div>
+    </React.Fragment>
+  );
+
+  const renderPlacedTooltip = () => (
+    <div
+      className="rw-tooltip-body"
+      ref={setPopperElement}
+      style={styles.popper}
+      {...attributes.popper}
+    >
+      {renderTooltipContent()}
+      <div
+        className="rw-tooltip-decoration rw-popper-arrow"
+        ref={setArrowElement}
+        style={styles.arrow}
+      />
+    </div>
+  );
+
+  const renderToolTip = () => (
+    // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+    <div className="rw-tooltip-body" style={{ backgroundColor: assistBackgoundColor }} onClick={(e) => { e.stopPropagation(); }}>
+      {renderTooltipContent()}
       <div className="rw-tooltip-decoration" style={{ backgroundColor: assistBackgoundColor }} />
     </div>
   );
@@ -78,7 +186,7 @@ const Launcher = ({
         <div className="rw-unread-count-pastille">{unreadCount}</div>
       )}
       <img src={openLauncherImage || openLauncher} className="rw-open-launcher" alt="" />
-      {showTooltip && lastMessage.get('sender') === 'response' && renderToolTip()}
+      {showTooltip && lastMessage && lastMessage.get('sender') === 'response' && (referenceElement ? renderPlacedTooltip() : renderToolTip())}
     </div>
   );
 
@@ -108,14 +216,39 @@ Launcher.propTypes = {
   unreadCount: PropTypes.number,
   displayUnreadCount: PropTypes.bool,
   showTooltip: PropTypes.bool,
-  lastMessage: ImmutablePropTypes.map
+  lastUserMessage: PropTypes.oneOfType([ImmutablePropTypes.map, PropTypes.bool]),
+  domHighlight: PropTypes.shape({}),
+  lastMessages: PropTypes.arrayOf(ImmutablePropTypes.map)
 };
 
 const mapStateToProps = state => ({
-  lastMessage: (state.messages && state.messages.get(-1)) || new Map(),
+  lastMessages: (state.messages && (() => {
+    const messages = [];
+    for (let i = 1; i <= 10; i += 1) {
+      if (state.messages.get(-i) && state.messages.get(-i).get('sender') !== 'response') break;
+      if (!state.messages.get(-i)) break;
+      messages.unshift(state.messages.get(-i));
+    }
+    return messages;
+  })()) || new Map(),
   unreadCount: state.behavior.get('unreadCount') || 0,
   showTooltip: state.metadata.get('showTooltip'),
-  linkTarget: state.metadata.get('linkTarget')
+  linkTarget: state.metadata.get('linkTarget'),
+  lastUserMessage: (function getLastUserMessage() {
+    if (!state.messages) return false;
+    let index = -1;
+    while (index > -10) {
+      const lastMessage = state.messages.get(index);
+      if (lastMessage) {
+        if (lastMessage.get('sender') === 'client') return lastMessage;
+      } else {
+        return false;
+      }
+      index -= 1;
+    }
+    return false;
+  }()),
+  domHighlight: state.metadata.get('domHighlight')
 });
 
 const mapDispatchToProps = dispatch => ({