+++ src/index.js
@@ -1,4 +1,4 @@
-import React, { forwardRef } from 'react';
+import React, { forwardRef, useRef, useEffect } from 'react';
 
 import PropTypes from 'prop-types';
 import { Provider } from 'react-redux';
@@ -82,28 +82,41 @@ const ConnectedWidget = forwardRef((props, ref) => {
     }
   }
 
-  const sock = new Socket(
-    props.socketUrl,
-    props.customData,
-    props.socketPath,
-    props.protocol,
-    props.protocolOptions,
-    props.onSocketEvent
-  );
+  useEffect(() => () => {
+    store = null;
+  }, []);
 
+  const instanceSocket = useRef({});
+
+  if (!instanceSocket.current.url && !(store && store.socketRef)) {
+    instanceSocket.current = new Socket(
+      props.socketUrl,
+      props.customData,
+      props.socketPath,
+      props.protocol,
+      props.protocolOptions,
+      props.onSocketEvent
+    );
+  }
+
+  if (!instanceSocket.current.url && store && store.socketRef) {
+    instanceSocket.current = store.socket;
+  }
 
   const storage =
     props.params.storage === 'session' ? sessionStorage : localStorage;
-  if (!store || sock.marker !== store.socketRef) {
+
+  if (!store) {
     store = initStore(
       props.inputTextFieldHint,
       props.connectingText,
-      sock,
+      instanceSocket.current,
       storage,
       props.docViewer,
       props.onWidgetEvent
     );
-    store.socketRef = sock.marker;
+    store.socketRef = instanceSocket.current.marker;
+    store.socket = instanceSocket.current;
   }
   return (
     <Provider store={store}>
@@ -137,7 +150,7 @@ const ConnectedWidget = forwardRef((props, ref) => {
           closeImage={props.closeImage}
           customComponent={props.customComponent}
           displayUnreadCount={props.displayUnreadCount}
-          socket={sock}
+          socket={instanceSocket.current}
           showMessageDate={props.showMessageDate}
           customMessageDelay={props.customMessageDelay}
           tooltipPayload={props.tooltipPayload}